Task 10

+++ pointers_arrays_strings/100-atoi.c
@@ -14,7 +14,7 @@
 int _atoi(char *s)
 {
 	int i;
-	int num;
+	int64_t num;
 	int a;
 	int status;
